Final before upload

+++ Asst1/mymalloc.h
@@ -6,7 +6,7 @@
 #define malloc( x ) mymalloc( x, __FILE__, __LINE__ )
 #define free( x ) myfree( x, __FILE__, __LINE__ )
 
-void *mymalloc(int size, char* file, int line);
+void *mymalloc(size_t size, char* file, int line);
 void myfree(void *ptr, char* file, int line);
 void listmem();
